Move chat page into src app directory
