--- app/page.tsx
@@ -1,5 +1,0 @@
-import PremiumCharacterChat from './PremiumCharacterChat';
-
-export default function Page() {
-  return <PremiumCharacterChat />;
-}
