+++ src/components/PremiumCharacterChat.tsx
